Refactor dialog and filter components
