+++ tnp-frontend/src/pages/Customer/CustomerList.jsx
@@ -94,354 +94,16 @@ import {
   open_dialog_error,
 } from "../../utils/import_lib";
 import ErrorBoundary from "../../components/ErrorBoundary";
-
-const StyledDataGrid = styled(DataGrid)(({ theme }) => ({
-  "& .MuiDataGrid-columnHeader": {
-    backgroundColor: theme.palette.error.dark,
-    color: theme.palette.common.white,
-    transition: "all 0.2s ease",
-    "&:hover": {
-      backgroundColor: theme.palette.error.main,
-    },
-  },
-
-  "& .MuiDataGrid-columnHeaderTitleContainer": {
-    justifyContent: "center",
-    fontSize: "0.95rem",
-    fontWeight: "bold",
-    letterSpacing: "0.5px",
-  },
-
-  "& .MuiDataGrid-row--borderBottom .MuiDataGrid-columnHeader": {
-    borderBottom: `1px solid ${theme.palette.error.dark}`,
-  },
-
-  "& .MuiDataGrid-columnHeader[aria-colindex='1']": {
-    borderBottomLeftRadius: theme.shape.borderRadius,
-  },
-
-  "& .MuiDataGrid-columnHeader--last": {
-    borderBottomRightRadius: theme.shape.borderRadius,
-  },
-
-  "& .MuiDataGrid-iconSeparator": {
-    display: "none",
-  },
-  "& .MuiDataGrid-row": {
-    backgroundColor: theme.vars.palette.grey.main,
-    borderRadius: theme.shape.borderRadius,
-    marginTop: 10,
-    transition: "all 0.2s ease",
-    "&:hover": {
-      backgroundColor: theme.vars.palette.grey.light,
-      transform: "translateY(-2px)",
-      boxShadow: "0 4px 8px rgba(0,0,0,0.08)",
-    },
-  },
-  "& .MuiDataGrid-cell, .MuiDataGrid-filler > div": {
-    textAlign: "center",
-    borderWidth: 0,
-    color: theme.vars.palette.grey.dark,
-    padding: "8px 16px",
-    display: "flex",
-    alignItems: "center",
-    fontSize: "0.95rem",
-    transition: "all 0.15s ease-in-out",
-    position: "relative",
-    "&:hover": {
-      transform: "scale(1.02)",
-    },
-    "&:focus": {
-      outline: "none",
-      backgroundColor: `${theme.palette.primary.light}11`,
-    },
-    "&::after": {
-      content: '""',
-      position: "absolute",
-      bottom: 0,
-      left: "10%",
-      width: "80%",
-      height: "1px",
-      backgroundColor: `${theme.palette.divider}`,
-      opacity: 0.5,
-    },
-  },
-
-  "& .MuiDataGrid-menuIcon > button > svg": {
-    color: "#fff",
-  },
-
-  "& .MuiDataGrid-iconButtonContainer > button > svg": {
-    color: "#fff",
-  },
-
-  "& .MuiDataGrid-actionsCell > .MuiIconButton-root:not(.Mui-disabled) > svg": {
-    color: theme.vars.palette.grey.dark,
-  },
-
-  "& .MuiDataGrid-footerContainer": {
-    borderWidth: 0,
-    justifyContent: "center",
-    padding: "16px 0",
-  },
-
-  "& .uppercase-cell": {
-    textTransform: "uppercase",
-  },
-  "& .danger-days": {
-    color: theme.vars.palette.error.main,
-    fontWeight: "bold",
-    padding: "4px 8px",
-    borderRadius: "12px",
-    backgroundColor: `${theme.palette.error.light}22`,
-  },
-
-  "& .warning-days": {
-    color: theme.vars.palette.warning.main,
-    fontWeight: "bold",
-    padding: "4px 8px",
-    borderRadius: "12px",
-    backgroundColor: `${theme.palette.warning.light}22`,
-  },
-  "& .MuiDataGrid-toolbarContainer": {
-    gap: 2,
-    padding: "12px 20px",
-    justifyContent: "space-between",
-    backgroundColor: theme.palette.error.dark,
-    borderTopLeftRadius: theme.shape.borderRadius,
-    borderTopRightRadius: theme.shape.borderRadius,
-    marginBottom: 10,
-    borderBottom: "2px solid rgba(255,255,255,0.1)",
-    boxShadow: "0 2px 8px rgba(0,0,0,0.15)",
-  },
-  // Animations for visual appeal
-  "@keyframes glow-border": {
-    "0%": {
-      boxShadow: `0 0 5px rgba(244, 67, 54, 0.5), inset 0 0 5px rgba(244, 67, 54, 0.1)`,
-    },
-    "50%": {
-      boxShadow: `0 0 20px rgba(244, 67, 54, 0.8), inset 0 0 10px rgba(244, 67, 54, 0.3)`,
-    },
-    "100%": {
-      boxShadow: `0 0 5px rgba(244, 67, 54, 0.5), inset 0 0 5px rgba(244, 67, 54, 0.1)`,
-    },
-  },
-
-  // Subtle pulse animation for attention
-  "@keyframes subtle-pulse": {
-    "0%, 100%": {
-      opacity: 1,
-      transform: "scale(1)",
-    },
-    "50%": {
-      opacity: 0.95,
-      transform: "scale(1.02)",
-    },
-  },
-
-  // Left indicator animation
-  "@keyframes slide-in-left": {
-    from: {
-      transform: "translateX(-100%)",
-      opacity: 0,
-    },
-    to: {
-      transform: "translateX(0)",
-      opacity: 1,
-    },
-  },
-  // Highlight rows based on recall days
-  "& .high-priority-row": {
-    backgroundColor: `${theme.palette.error.light}33`, // 20% opacity red for <= 7 days
-    animation: "glow-border 2s ease-in-out infinite",
-    border: `2px solid ${theme.palette.error.main}`,
-    borderRadius: theme.shape.borderRadius,
-    position: "relative",
-    zIndex: 1,
-    "&:hover": {
-      backgroundColor: `${theme.palette.error.light}66`, // 40% opacity
-      transform: "translateY(-3px)",
-      boxShadow: `0 6px 12px ${theme.palette.error.light}66`,
-    },
-    "&::before": {
-      content: '""',
-      position: "absolute",
-      top: 0,
-      left: -8,
-      height: "100%",
-      width: "4px",
-      backgroundColor: theme.palette.error.main,
-      borderRadius: "2px",
-    },
-  },
-
-  "& .medium-priority-row": {
-    backgroundColor: `${theme.palette.warning.light}33`, // 20% opacity yellow for <= 15 days
-    border: `1px solid ${theme.palette.warning.main}66`,
-    borderRadius: theme.shape.borderRadius,
-    position: "relative",
-    zIndex: 0,
-    "&:hover": {
-      backgroundColor: `${theme.palette.warning.light}66`, // 40% opacity
-      transform: "translateY(-2px)",
-      boxShadow: `0 4px 10px ${theme.palette.warning.light}66`,
-    },
-    "&::before": {
-      content: '""',
-      position: "absolute",
-      top: 0,
-      left: -8,
-      height: "100%",
-      width: "4px",
-      backgroundColor: theme.palette.warning.main,
-      borderRadius: "2px",
-    },
-  },
-}));
-
-const StyledPagination = styled(Pagination)(({ theme }) => ({
-  "& .MuiPaginationItem-previousNext": {
-    backgroundColor: theme.vars.palette.error.dark,
-    color: "#fff",
-    height: 30,
-    width: 38,
-
-    "&:hover": {
-      backgroundColor: theme.vars.palette.error.main,
-    },
-  },
-
-  "& .MuiPaginationItem-page": {
-    backgroundColor: theme.vars.palette.grey.outlinedInput,
-    borderColor: theme.vars.palette.grey.outlinedInput,
-    height: 30,
-    width: 38,
-
-    "&:hover": {
-      backgroundColor: theme.vars.palette.grey.light,
-      borderColor: theme.vars.palette.grey.light,
-    },
-  },
-
-  "& .MuiPaginationItem-ellipsis": {
-    backgroundColor: theme.vars.palette.grey.outlinedInput,
-    borderColor: theme.vars.palette.grey.outlinedInput,
-    borderRadius: theme.vars.shape.borderRadius,
-    height: 30,
-    width: 38,
-    alignContent: "center",
-  },
-  "& .MuiPaginationItem-page.Mui-selected": {
-    backgroundColor: theme.vars.palette.error.light,
-    borderColor: theme.vars.palette.error.light,
-    color: theme.palette.common.white,
-    fontWeight: "bold",
-    transform: "scale(1.05)",
-    boxShadow: "0 2px 4px rgba(0,0,0,0.2)",
-    transition: "all 0.2s ease",
-
-    "&:hover": {
-      backgroundColor: theme.vars.palette.error.main,
-    },
-  },
-}));
-
-// Custom component for page size selection
-const PageSizeSelector = ({ value, onChange }) => {
-  const pageSizeOptions = [30, 50, 80, 100];
-
-  return (
-    <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-      <Typography
-        variant="body2"
-        sx={{ color: (theme) => theme.vars.palette.grey.dark }}
-      >
-        Rows per page:
-      </Typography>
-      <FormControl size="small" sx={{ minWidth: 85 }}>
-        <Select
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          displayEmpty
-          variant="outlined"
-          sx={{
-            borderRadius: 1,
-            backgroundColor: (theme) => theme.vars.palette.grey.outlinedInput,
-            ".MuiOutlinedInput-notchedOutline": { borderColor: "transparent" },
-            "&:hover .MuiOutlinedInput-notchedOutline": {
-              borderColor: "transparent",
-            },
-            "&.Mui-focused .MuiOutlinedInput-notchedOutline": {
-              borderColor: "transparent",
-            },
-            ".MuiSelect-select": { py: 0.5, px: 1 },
-          }}
-        >
-          {pageSizeOptions.map((option) => (
-            <MenuItem key={option} value={option}>
-              {option} rows
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-    </Box>
-  );
-};
+import { StyledDataGrid, StyledPagination } from "./components/StyledComponents";
+import PageSizeSelector from "./components/PageSizeSelector";
+import SortInfoDisplay from "./components/SortInfoDisplay";
+import CustomPagination from "./components/CustomPagination";
+import CustomToolbar from "./components/CustomToolbar";
 
 const channelMap = {
   1: "sales",
   2: "online",
   3: "office",
-};
-
-// Custom component for information about column sorting
-const SortInfoDisplay = ({ sortModel }) => {
-  if (!sortModel || sortModel.length === 0) {
-    return null;
-  }
-  const fieldMap = {
-    cus_no: "รหัสลูกค้า",
-    cus_channel: "ช่องทาง",
-    cus_bt_id: "ประเภทธุรกิจ",
-    business_type: "ประเภทธุรกิจ", // Keep for backward compatibility
-    cus_manage_by: "ชื่อเซลล์",
-    cus_name: "ชื่อลูกค้า",
-    cus_company: "ชื่อบริษัท",
-    cus_tel_1: "เบอร์โทร",
-    cd_last_datetime: "วันติดต่อกลับ",
-    cd_note: "หมายเหตุ",
-    cus_email: "อีเมล",
-    cus_address: "ที่อยู่",
-  };
-
-  const { field, sort } = sortModel[0];
-  const displayField = fieldMap[field] || field;
-  const displayDirection = sort === "asc" ? "ascending" : "descending";
-
-  // Use icons to make it more visually clear
-  const SortIcon =
-    sort === "asc"
-      ? () => <span style={{ fontSize: "0.8em", marginRight: "4px" }}>▲</span>
-      : () => <span style={{ fontSize: "0.8em", marginRight: "4px" }}>▼</span>;
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        alignItems: "center",
-        gap: 1,
-        padding: "4px 8px",
-        borderRadius: "4px",
-        backgroundColor: "rgba(255, 255, 255, 0.2)",
-        color: "white",
-      }}
-    >
-      <SortIcon />
-      <Typography variant="caption" sx={{ fontWeight: "medium" }}>
-        เรียงตาม: {displayField} (
-        {displayDirection === "ascending" ? "น้อยไปมาก" : "มากไปน้อย"})
-      </Typography>
-    </Box>
-  );
 };
 
 function CustomerList() {
@@ -642,94 +304,6 @@ function CustomerList() {
       console.warn("Failed to save column order to localStorage", error);
     }
   }; // Pagination customize
-  function CustomPagination() {
-    const apiRef = useGridApiContext();
-    const page = useGridSelector(apiRef, gridPageSelector);
-    const pageCount = useGridSelector(apiRef, gridPageCountSelector);
-    const theme = useTheme();
-    const isXs = useMediaQuery(theme.breakpoints.down("sm"));
-
-    // Reset page to first page after change group.
-    useEffect(() => {
-      if (paginationModel.page !== page) {
-        apiRef.current.setPage(0);
-        scrollToTop();
-      }
-    }, [paginationModel, scrollToTop]);
-
-    // Handle page size change
-    const handlePageSizeChange = (newPageSize) => {
-      const newModel = { ...paginationModel, pageSize: newPageSize, page: 0 };
-      dispatch(setPaginationModel(newModel));
-      apiRef.current.setPageSize(newPageSize);
-      scrollToTop();
-    };
-    return (
-      <Box
-        sx={{
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "space-between",
-          flexWrap: "wrap",
-          gap: 2,
-          width: "100%",
-          p: 1,
-        }}
-      >
-        <PageSizeSelector
-          value={paginationModel.pageSize}
-          onChange={handlePageSizeChange}
-        />
-
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            gap: 2,
-            flexWrap: "wrap",
-            justifyContent: "center",
-            flex: 1,
-          }}
-        >
-          <StyledPagination
-            color="error"
-            variant="outlined"
-            shape="rounded"
-            page={page + 1}
-            count={pageCount}
-            siblingCount={isXs ? 0 : 1}
-            boundaryCount={1}
-            // @ts-expect-error
-            renderItem={(props2) => (
-              <PaginationItem
-                {...props2}
-                disableRipple
-                slots={{ previous: FaChevronLeft, next: FaChevronRight }}
-              />
-            )}
-            onChange={(event, value) => {
-              apiRef.current.setPage(value - 1);
-              scrollToTop();
-            }}
-          />
-        </Box>
-
-        <Typography
-          variant="body2"
-          sx={{
-            color: (theme) => theme.vars.palette.grey.dark,
-            minWidth: 120,
-            textAlign: "right",
-          }}
-        >
-          {`${page * paginationModel.pageSize + 1}-${Math.min(
-            (page + 1) * paginationModel.pageSize,
-            totalItems
-          )} of ${totalItems}`}
-        </Typography>
-      </Box>
-    );
-  }
   const handleOpenDialog = (mode, cus_id = null) => {
     // First, reset the input list to avoid any stale data
     dispatch(resetInputList());
@@ -1678,48 +1252,6 @@ function CustomerList() {
     ]
   );
 
-  // Custom toolbar component
-  function CustomToolbar() {
-    return (
-      <GridToolbarContainer>
-        <Box sx={{ display: "flex", alignItems: "center", flex: 1 }}>
-          <Typography
-            variant="subtitle2"
-            sx={{
-              color: "common.white",
-              fontWeight: "bold",
-              display: "flex",
-              alignItems: "center",
-              mr: 2,
-            }}
-          >
-            รายการลูกค้า
-          </Typography>{" "}
-          <SortInfoDisplay sortModel={serverSortModel} />
-        </Box>{" "}
-        <Box sx={{ display: "flex", gap: 1, alignItems: "center" }}>
-          {isFetching && (
-            <Box
-              sx={{
-                display: "flex",
-                alignItems: "center",
-                marginRight: 1,
-                color: "white",
-                fontSize: "0.75rem",
-                backgroundColor: "rgba(255,255,255,0.2)",
-                padding: "4px 8px",
-                borderRadius: "4px",
-                gap: 1,
-              }}
-            >
-              <CircularProgress size={16} thickness={5} color="inherit" />
-              <Typography variant="caption">กำลังโหลด...</Typography>
-            </Box>          )}
-          {/* ColumnVisibilitySelector component removed as requested */}
-        </Box>
-      </GridToolbarContainer>
-    );
-  }
 
   return (
     <ScrollContext.Provider value={{ scrollToTop }}>
@@ -1824,6 +1356,10 @@ function CustomerList() {
                 noRowsOverlay: NoDataComponent,
                 pagination: CustomPagination,
                 toolbar: CustomToolbar,
+              }}
+              slotProps={{
+                pagination: { paginationModel, scrollToTop, totalItems, dispatch },
+                toolbar: { serverSortModel, isFetching },
               }}
               sx={{ border: 0 }}
               rowHeight={60}
