+++ tnp-frontend/src/pages/Customer/DialogForm.jsx
@@ -8,10 +8,7 @@ import {
   DialogTitle,
   DialogContent,
   DialogActions,
-  OutlinedInput,
   Grid2 as Grid,
-  styled,
-  Select,
   MenuItem,
   InputLabel,
   IconButton,
@@ -19,7 +16,6 @@ import {
   Tooltip,
   Paper,
   Typography,
-  TextField,
   Tabs,
   Tab,
   Divider,
@@ -71,74 +67,14 @@ import {
   open_dialog_loading,
 } from "../../utils/import_lib";
 import Swal from "sweetalert2";
-
-// Custom styled components
-const StyledTextField = styled(TextField)(({ theme }) => ({
-  "& .MuiOutlinedInput-root": {
-    backgroundColor: theme.vars.palette.grey.outlinedInput,
-    "&.Mui-disabled": {
-      "& .MuiOutlinedInput-notchedOutline": {
-        borderColor: theme.vars.palette.grey.outlinedInput,
-      },
-      "& .MuiOutlinedInput-input": {
-        WebkitTextFillColor: theme.vars.palette.text.primary,
-      },
-    },
-  },
-  "& .MuiInputLabel-root": {
-    color: theme.vars.palette.grey.dark,
-    fontFamily: "Kanit",
-    fontSize: 14,
-  },
-}));
-
-const StyledSelect = styled(Select)(({ theme }) => ({
-  backgroundColor: theme.vars.palette.grey.outlinedInput,
-  "& fieldset": {
-    borderColor: theme.vars.palette.grey.outlinedInput,
-  },
-}));
-
-const SectionTitle = styled(Typography)(({ theme }) => ({
-  fontWeight: 500,
-  marginBottom: theme.spacing(1),
-  color: theme.vars.palette.primary.main,
-  display: "flex",
-  alignItems: "center",
-  gap: theme.spacing(1),
-}));
-
-const FormSection = styled(Box)(({ theme }) => ({
-  marginBottom: theme.spacing(2),
-  padding: theme.spacing(2),
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: theme.palette.background.paper,
-  boxShadow: theme.shadows[1],
-}));
-
-// Tab Panel component for tabbed interface
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`customer-tabpanel-${index}`}
-      aria-labelledby={`customer-tab-${index}`}
-      {...other}
-    >
-      {value === index && <Box sx={{ py: 2 }}>{children}</Box>}
-    </div>
-  );
-}
-
-function a11yProps(index) {
-  return {
-    id: `customer-tab-${index}`,
-    "aria-controls": `customer-tabpanel-${index}`,
-  };
-}
+import {
+  StyledTextField,
+  StyledSelect,
+  SectionTitle,
+  FormSection,
+  TabPanel,
+  a11yProps,
+} from "./components/DialogFormStyled";
 
 function DialogForm(props) {
   const dispatch = useDispatch();
